// frontend/src/shared/types/api.ts
export type Payment = {
  id: string;
  order_id: string;
  user_id: string;
  order_total: number;
  amount: number;
  status: string;
  transaction_type: string;
  reference_payment_id?: string;
  payment_method: string;
  gateway_provider: string;
  gateway_transaction_id?: string;
  gateway_order_id?: string;
  checkout_url?: string;
  signature_verified: boolean;
  failure_reason?: string;
  net_paid_amount?: number;
  outstanding_amount?: number;
  created_at: string;
  updated_at: string;
};